Moved resolution logic into reference class
This means that now everything which can be a reference always has a
reference, even if it's a reference which resolves to null. This is
important for completion, as PsiReference#getVariants() is used to find
the list of completions.
