+++ src/glslplugin/lang/elements/expressions/GLSLIdentifierExpression.java
@@ -100,75 +100,6 @@ public class GLSLIdentifierExpression extends GLSLExpression implements GLSLRefe
     @Override
     @Nullable
     public GLSLVariableReference getReferenceProxy() {
-        GLSLDeclarator target = getVariableReference(this);
-
-        if (target != null) {
-            return new GLSLVariableReference(this.getNameIdentifier(), target);
-        } else {
-            return null;
-        }
+        return new GLSLVariableReference(this);
     }
-
-    @Nullable
-    private GLSLDeclarator getVariableReference(PsiElement start) {
-        PsiElement current = start.getPrevSibling();
-        GLSLDeclarator result = null;
-        if (current == null) {
-            current = start.getParent();
-        }
-
-        while (current != null) {
-
-            // Only process it if we haven't already done so.
-            if (current instanceof GLSLDeclarationList && !isDescendantOf(current)) {
-                GLSLDeclarationList list = (GLSLDeclarationList) current;
-                for (GLSLDeclaration declaration : list.getDeclarations()) {
-                    result = getVariableReferenceCheckDeclaration(declaration);
-                    if (result != null) {
-                        break;
-                    }
-                }
-            } else {
-                GLSLVariableDeclaration declaration = null;
-
-                if (current instanceof GLSLDeclarationStatement) {
-                    declaration = ((GLSLDeclarationStatement) current).getDeclaration();
-                }
-
-                if (current instanceof GLSLVariableDeclaration) {
-                    declaration = (GLSLVariableDeclaration) current;
-                }
-
-                if (declaration != null) {
-                    result = getVariableReferenceCheckDeclaration(declaration);
-                }
-            }
-
-            if (result != null) {
-                return result;
-            }
-
-            if (current.getPrevSibling() == null) {
-                current = current.getParent();
-                if (current instanceof GLSLFile) {
-                    current = null;
-                }
-            } else {
-                current = current.getPrevSibling();
-            }
-        }
-
-        return null;
-    }
-
-    @Nullable
-    private GLSLDeclarator getVariableReferenceCheckDeclaration(GLSLDeclaration declaration) {
-        for (GLSLDeclarator declarator : declaration.getDeclarators()) {
-            if (declarator.getName().equals(getName())) {
-                return declarator;
-            }
-        }
-        return null;
-    }
-
 }
